Fix modal styling, add separate components for different content stages
